refactor: separate settings pane lifecycle
